Struct uic_angel::cache::Cache

pub struct Cache {
    groups_for_endpoints: BTreeMap<(String, u16), Vec<u16>>,
    reverse_lookup: BTreeMap<u16, BTreeMap<String, Vec<u16>>>,
    group_names: BTreeMap<u16, String>,
    supported_commands_for_endpoints: BTreeMap<(String, u16), BTreeMap<String, HashSet<String>>>,
}

Fields
groups_for_endpoints: BTreeMap<(String, u16), Vec<u16>>reverse_lookup: BTreeMap<u16, BTreeMap<String, Vec<u16>>>group_names: BTreeMap<u16, String>supported_commands_for_endpoints: BTreeMap<(String, u16), BTreeMap<String, HashSet<String>>>
Implementations

impl Cache

pub fn new() -> Self

pub fn set_group_list_for_node(
&mut self,
unid: &str,
endpoint: u16,
groups: &[u16]
) -> Vec<EntryChanged>

Update the cache with new groups for an given key (unid-endpoint combination) if unid.is_empty(), the function returns and does not update anything. returns a list with changed endpoints for all effected keys. if the endpoint vector inside EntryChanged is empty, its considered unretained

pub fn get_group_list_for_node(&mut self, unid: &str, endpoint: u16) -> Vec<u16>

Get a list of groups a node belongs to. Returns an empty vector if unid is empty or if node doesn’t belong to a group.

fn make_changed_entries(
&self,
key: (String, u16),
added_groups: &[u16],
removed_groups: &[u16]
) -> Vec<EntryChanged>

creates EntryChanged objects for the records that will be effected by the incoming change. this function assumes the groups_for_endpoints already reflects the new state. if the endpoint vector inside EntryChanged is empty, its considered unretained

fn update_reverse_lookup(&mut self)

rebuilds reverse_lookup completely from groups_for_endpoints

pub fn update_group_name(&mut self, groupid: u16, name: &str) -> bool

returns true if the name was updated. returns false when trying to update an existing name to an empty string or when the new name == the old name

pub fn get_group_name(&self, groupid: u16) -> Option<String>

pub fn remove_group_name(&mut self, groupid: u16)

pub fn set_endpoint_cluster_supported_commands(
&mut self,
key: (String, u16),
cluster_name: &str,
commands: HashSet<String>
)

Update the list of supported-commands, per cluster, for the given endpoint.

pub fn get_endpoint_cluster_supported_commands(
&self,
key: &(String, u16),
cluster_name: &str
) -> Option<HashSet<String>>

Get the list of supported commands, for a given cluster, for a given endpoint.

pub fn get_endpoints_for_group(
&self,
group_id: &u16
) -> Option<Vec<(String, u16)>>

Get the list of endpoints that belong to group.

pub fn get_group_supported_commands_per_cluster(
&self,
group_id: &u16
) -> Option<BTreeMap<String, HashSet<String>>>

Get the list of commands supported by group per clusters.

pub fn get_endpoints_supported_commands_per_cluster(
&self,
end_points: Vec<(String, u16)>
) -> Option<BTreeMap<String, HashSet<String>>>

Get the list of commands supported by endpoint per clusters.