Struct uic_angel::mqtt_handler::MqttHandler

pub struct MqttHandler<'a> {
    mqtt_client: Rc<dyn MqttClient<'a>>,
    cache: Cache,
    reported_group_list_matcher: TopicMatcher<'a>,
    reported_names_matcher: TopicMatcher<'a>,
    supported_commands: TopicMatcher<'a>,
}

MqttHandler handles subscription messages related to UIC groups. It subscribes to listen to the correct topics and is to be passed to the mosquitto client

Fields
mqtt_client: Rc<dyn MqttClient<'a>>

Reference counted to the mosquitto-client. Notice that its lifetime spans as long as mqttHandler itself. e.g. the mqtt_client is locked for immutablilty as long as the MqttHandler exists.

cache: Cache

Lookup cache for groups and unids

reported_group_list_matcher: TopicMatcher<'a>reported_names_matcher: TopicMatcher<'a>supported_commands: TopicMatcher<'a>
Implementations

impl<'a> MqttHandler<'a>

pub fn new(
client: Rc<dyn MqttClient<'a>>,
matchers: (TopicMatcher<'a>, TopicMatcher<'a>, TopicMatcher<'a>)
) -> Self

crate new handler to be passed into a MqttClient. The native mqtt client is passed as reference counted. Reference counted objects can be shared and borrowed over multiple instances. Acquirring mutablity is only possible when the ref count == 1. since this rc is shared between MqttClient and MqttHandler, reference count can never be == 1. hence mutable cannot be accuired for mqtt client.

fn process_message(&mut self, msg: MosqMessage) -> Result<(), String>

fn get_unid_ep_of_topic(topic: &str) -> Result<(String, u16), String>

parse group topic message and extract unid and endpoint id

fn on_group_list_reported(
&mut self,
topic: &str,
json: JsonValue
) -> Result<(), String>

process a ucl/by-unid/+/+/Groups/Attributes/GroupList/Reported topic stores a list of groups for an unid and publishes its endpoints on the ucl/by-group//NodeList/ expects endpoint to be prefixed with ep. e.g. ep0

fn publish_commands_by_cluster(
&mut self,
group_id: &u16,
supported_commands: BTreeMap<String, HashSet<String>>,
is_unretain: bool
) -> Result<(), String>

fn on_name_reported(
&mut self,
topic: &str,
json: JsonValue
) -> Result<(), String>

handler for topic ucl/by-unid/+/+/Groups/Attributes/+/Name/Reported. overwrites name for group id.

fn send_add_group_if_required(
&self,
group_id: u16,
group_name: &str
) -> Result<(), String>

add_group is only send when an update of an groupname is reported

fn send_group_name(&self, group_id: u16, group_name: &str) -> Result<(), String>

fn on_supported_commands(
&mut self,
topic: &str,
json: JsonValue
) -> Result<(), String>